requirements.txt
requests
pyyaml
keyboard
playwright
